+++ src/GenericContainer/CMakeLists.txt
@@ -5,7 +5,7 @@
 # set target name
 add_bipedal_component(
     NAME                   GenericContainer
-    PUBLIC_HEADERS         include/BipedalLocomotionControllers/GenericContainer.h include/BipedalLocomotionControllers/TemplateHelpers.h
+    PUBLIC_HEADERS         include/BipedalLocomotionControllers/GenericContainer/Vector.h include/BipedalLocomotionControllers/GenericVector/TemplateHelpers.h
     PUBLIC_LINK_LIBRARIES ${iDynTree_LIBRARIES}
     SUBDIRECTORIES   tests
     IS_INTERFACE     ON)
